Split services into their own modules
