+++ bitsv/network/services/__init__.py
@@ -1,0 +1,7 @@
+from .bitindex import BitIndex
+from .bitindex3 import (
+    BitIndex3, BitIndex3MainNet, BitIndex3TestNet, BitIndex3STN
+)
+from .insight import InsightAPI
+from .network import NetworkAPI
+from .whatsonchain import WhatsonchainAPI
