+++ bitsv/network/services/network.py
@@ -1,0 +1,139 @@
+import requests
+
+from .bitindex import BitIndex
+
+
+class NetworkAPI:
+    """
+    A Class for handling network API redundancy.
+
+    FIXME network API redundancy and RegTesting / Testnet - see Issue section on github"""
+
+    IGNORED_ERRORS = (ConnectionError,
+                      requests.exceptions.ConnectionError,
+                      requests.exceptions.Timeout,
+                      requests.exceptions.ReadTimeout)
+
+    # BitIndex for all mainnet functions
+    GET_BALANCE_MAIN = [BitIndex.get_balance]
+    GET_TRANSACTIONS_MAIN = [BitIndex.get_txs]
+    GET_UNSPENT_MAIN = [BitIndex.get_utxo]
+    BROADCAST_TX_MAIN = [BitIndex.broadcast_rawtx]
+    GET_TX_MAIN = [BitIndex.get_tx]
+
+    @classmethod
+    def get_balance(cls, address):
+        """Gets the balance of an address in satoshis.
+
+        :param address: The address in question.
+        :type address: ``str``
+        :raises ConnectionError: If all API services fail.
+        :rtype: ``int``
+        """
+
+        for api_call in cls.GET_BALANCE_MAIN:
+            try:
+                return api_call(address)
+            except cls.IGNORED_ERRORS:
+                pass
+
+        raise ConnectionError('All APIs are unreachable.')
+
+    @classmethod
+    def get_transactions(cls, address):
+        """Gets the ID of all transactions related to an address.
+
+        :param address: The address in question.
+        :type address: ``str``
+        :raises ConnectionError: If all API services fail.
+        :rtype: ``list`` of ``str``
+        """
+
+        for api_call in cls.GET_TRANSACTIONS_MAIN:
+            try:
+                return api_call(address)
+            except cls.IGNORED_ERRORS:
+                pass
+
+        raise ConnectionError('All APIs are unreachable.')
+
+    @classmethod
+    def get_transaction(cls, txid):
+        """Gets the full transaction details.
+
+        :param txid: The transaction id in question.
+        :type txid: ``str``
+        :raises ConnectionError: If all API services fail.
+        :rtype: ``Transaction``
+        """
+
+        for api_call in cls.GET_TX_MAIN:
+            try:
+                return api_call(txid)
+            except cls.IGNORED_ERRORS:
+                pass
+
+        raise ConnectionError('All APIs are unreachable.')
+
+    @classmethod
+    def get_tx_amount(cls, txid, txindex):
+        """Gets the amount of a given transaction output.
+
+        :param txid: The transaction id in question.
+        :type txid: ``str``
+        :param txindex: The transaction index in question.
+        :type txindex: ``int``
+        :raises ConnectionError: If all API services fail.
+        :rtype: ``Decimal``
+        """
+
+        for api_call in cls.GET_TX_AMOUNT_MAIN:
+            try:
+                return api_call(txid, txindex)
+            except cls.IGNORED_ERRORS:
+                pass
+
+        raise ConnectionError('All APIs are unreachable.')
+
+    @classmethod
+    def get_unspent(cls, address):
+        """Gets all unspent transaction outputs belonging to an address.
+
+        :param address: The address in question.
+        :type address: ``str``
+        :raises ConnectionError: If all API services fail.
+        :rtype: ``list`` of :class:`~bitcash.network.meta.Unspent`
+        """
+
+        for api_call in cls.GET_UNSPENT_MAIN:
+            try:
+                return api_call(address)
+            except cls.IGNORED_ERRORS:
+                pass
+
+        raise ConnectionError('All APIs are unreachable.')
+
+    @classmethod
+    def broadcast_tx(cls, tx_hex):  # pragma: no cover
+        """Broadcasts a transaction to the blockchain.
+
+        :param tx_hex: A signed transaction in hex form.
+        :type tx_hex: ``str``
+        :raises ConnectionError: If all API services fail.
+        """
+        success = None
+
+        for api_call in cls.BROADCAST_TX_MAIN:
+            try:
+                success = api_call(tx_hex)
+                if not success:
+                    continue
+                return
+            except cls.IGNORED_ERRORS:
+                pass
+
+        if success is False:
+            raise ConnectionError('Transaction broadcast failed, or '
+                                  'Unspents were already used.')
+
+        raise ConnectionError('All APIs are unreachable.')
